refactor: split thematic analysis into 2 files
